restructured the code
