--- filter_json_tables.py
@@ -1,27 +1,0 @@
-import json
-import os
-
-# Get A list of all files generated at the input directory
-def get_list_json_tables(tables_location):
-    #List of all the files generated by camelot
-    tableList = []
-    for root, dirs, files in os.walk(tables_location):
-        for file in files:
-            toAppend = os.path.join(tables_location,file).replace('\\', '/')
-            tableList.append(toAppend)
-
-    return tableList
-
-
-def find_exact_table(tableList):
-    for table in tableList:
-        jsonFile = json.load(open(table))
-        
-        #TODO
-        #Replace this if with a nicer thing, maybe add a list of all the needed values; this problem will solve itself with improved accuraty (using GhostScraper)
-        if 'ENVIRONMENTAL IMPACTS PER KG' in str(jsonFile) or 'RESOURCE USE PER DECLARED UNIT' in str(jsonFile) or 'Miljøpåvirkninger, 15 cm tyk væg' in str(jsonFile):
-            print(f'! Found the table at {table} !')
-            #TODO, get the exact table location by splitting the table after page - <> and table - <> 
-            #will greatly improve accuracy
-            #table_location = table.split("word",1)
-            return table
